requirements.txt
streamlit>=1.31.0
pandas>=2.2.0
plotly>=5.18.0
openai>=1.12.0
python-dotenv>=1.0.0
python-dateutil>=2.8.2 